+++ mpi_openmp/src/kmeans.c
@@ -8,7 +8,7 @@ void elementIntArray(int *array, int value, int size) {
     }
 }
 
-void elementFloatArray(int *array, float value, int size) {
+void elementFloatArray(float *array, float value, int size) {
     #pragma omp parallel for
     for(int i = 0; i < size; i++) {
         array[i] = value;
